// stylishlab-bridebox-frontend/src/api/generated/model/createSaleRequestPaymentStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Stylish Lab Bridebox - Salon Management API
 * API for managing salon services, employees, customers, sales, expenses, and reports
 * OpenAPI spec version: 1.0.0
 */

export type CreateSaleRequestPaymentStatus = typeof CreateSaleRequestPaymentStatus[keyof typeof CreateSaleRequestPaymentStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateSaleRequestPaymentStatus = {
  FULLY_PAID: 'FULLY_PAID',
  CREDIT: 'CREDIT',
  PARTIAL: 'PARTIAL',
} as const;